Move EventHorizon to its own project
