+++ Runtime.sln
@@ -63,6 +63,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Events.Store.MongoDB", "Sou
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Events.Store.MongoDB", "Specifications\Events.Store.MongoDB\Events.Store.MongoDB.csproj", "{129A7914-E5CE-41DC-BE0E-414C1D3F7FFE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventHorizon", "Source\EventHorizon\EventHorizon.csproj", "{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -412,6 +414,18 @@ Global
 		{129A7914-E5CE-41DC-BE0E-414C1D3F7FFE}.Release|x64.Build.0 = Release|Any CPU
 		{129A7914-E5CE-41DC-BE0E-414C1D3F7FFE}.Release|x86.ActiveCfg = Release|Any CPU
 		{129A7914-E5CE-41DC-BE0E-414C1D3F7FFE}.Release|x86.Build.0 = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|x64.Build.0 = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Debug|x86.Build.0 = Debug|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|x64.ActiveCfg = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|x64.Build.0 = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|x86.ActiveCfg = Release|Any CPU
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{AF810B2D-D506-4BAD-A61D-9A894CFC9990} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
@@ -442,5 +456,6 @@ Global
 		{CBC8298F-3F51-4AB3-AD5E-5F8818C1CBBF} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
 		{40D79C2E-B1D4-41FE-B1AF-F7176CEA9D15} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
 		{129A7914-E5CE-41DC-BE0E-414C1D3F7FFE} = {6DECF56C-D430-4E51-AB9A-59FD49309EEE}
+		{D3EA1FA0-D395-4EA0-8B0F-5335DE8218DB} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
 	EndGlobalSection
 EndGlobal
