+++ Source/EventHorizon/Consumer/EventHorizonClient.cs
@@ -20,7 +20,7 @@ using Dolittle.Runtime.Events.Store;
 using Dolittle.Tenancy;
 using grpc = contracts::Dolittle.Runtime.Events.Processing;
 
-namespace Dolittle.Runtime.Events.Processing.EventHorizon
+namespace Dolittle.Runtime.EventHorizon
 {
     /// <summary>
     /// Represents an implementation of <see cref="IEventHorizonClient" />.
