+++ Source/EventHorizon/EventHorizon.csproj
@@ -1,0 +1,26 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <Import Project="../../default.props"/>
+
+    <PropertyGroup>
+        <AssemblyName>Dolittle.Runtime.EventHorizon</AssemblyName>
+        <RootNamespace>Dolittle.Runtime.EventHorizon</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Dolittle.Concepts" Version="$(FundamentalsVersion)" />
+        <PackageReference Include="Dolittle.DependencyInversion" Version="$(FundamentalsVersion)" />
+        <PackageReference Include="Dolittle.Protobuf" Version="$(FundamentalsVersion)" />
+        <PackageReference Include="Dolittle.Runtime.Contracts" Version="$(ContractsVersion)" Alias="contracts" />
+        <PackageReference Include="Dolittle.Resilience" Version="$(FundamentalsVersion)" />
+        <PackageReference Include="Dolittle.Services.Clients" Version="$(FundamentalsVersion)" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Applications.Configuration\Applications.Configuration.csproj" />
+        <ProjectReference Include="..\Events.Streams\Events.Streams.csproj" />
+        <ProjectReference Include="..\Events.Processing\Events.Processing.csproj" />
+        <ProjectReference Include="..\Services\Services.csproj" />
+        <ProjectReference Include="..\Events.Store\Events.Store.csproj" />
+        <ProjectReference Include="..\Tenancy\Tenancy.csproj" />
+    </ItemGroup>
+</Project>
