+++ Source/EventHorizon/Producer/IEventHorizonSubscribers.cs
@@ -5,7 +5,7 @@ using Dolittle.Applications;
 using Dolittle.Runtime.Events.Streams;
 using Dolittle.Tenancy;
 
-namespace Dolittle.Runtime.Events.Processing.EventHorizon
+namespace Dolittle.Runtime.EventHorizon
 {
     /// <summary>
     /// Defines a system that knows about subscribers on the event horizon.
