+++ Source/EventHorizon/RuntimeServices.cs
@@ -4,12 +4,13 @@
 extern alias contracts;
 
 using System.Collections.Generic;
+using Dolittle.Runtime.Events.Processing;
 using Dolittle.Runtime.Events.Processing.Filters;
 using Dolittle.Runtime.Services;
 using Dolittle.Services;
 using grpc = contracts::Dolittle.Runtime.Events.Processing;
 
-namespace Dolittle.Runtime.Events.Processing
+namespace Dolittle.Runtime.EventHorizon
 {
     /// <summary>
     /// Represents an implementation of <see cref="ICanBindRuntimeServices"/> for exposing
@@ -19,6 +20,7 @@ namespace Dolittle.Runtime.Events.Processing
     {
         readonly FiltersService _filtersService;
         readonly EventHandlersService _eventHandlers;
+        readonly EventHorizonService _eventHorizon;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="RuntimeServices"/> class.
@@ -26,10 +28,11 @@ namespace Dolittle.Runtime.Events.Processing
         /// <param name="filtersService">The <see cref="FiltersService"/>.</param>
         /// <param name="eventHandlers">The <see cref="EventHandlersService"/>.</param>
         /// <param name="eventHorizon">The <see cref="EventHorizonService" />.</param>
-        public RuntimeServices(FiltersService filtersService, EventHandlersService eventHandlers)
+        public RuntimeServices(FiltersService filtersService, EventHandlersService eventHandlers, EventHorizonService eventHorizon)
         {
             _filtersService = filtersService;
             _eventHandlers = eventHandlers;
+            _eventHorizon = eventHorizon;
         }
 
         /// <inheritdoc/>
@@ -42,6 +45,7 @@ namespace Dolittle.Runtime.Events.Processing
             {
                 new Service(_filtersService, grpc.Filters.BindService(_filtersService), grpc.Filters.Descriptor),
                 new Service(_eventHandlers, grpc.EventHandlers.BindService(_eventHandlers), grpc.EventHandlers.Descriptor),
+                new Service(_eventHorizon, grpc.EventHorizon.BindService(_eventHorizon), grpc.EventHorizon.Descriptor)
             };
         }
     }
